refactor: [#1187] move fn initialize_tracker_dependencies
