+++ src/core/services/statistics/mod.rs
@@ -117,7 +117,7 @@ mod tests {
     use torrust_tracker_primitives::torrent_metrics::TorrentsMetrics;
     use torrust_tracker_test_helpers::configuration;
 
-    use crate::bootstrap::app::initialize_tracker_dependencies;
+    use crate::app_test::initialize_tracker_dependencies;
     use crate::core;
     use crate::core::services::statistics::{self, get_metrics, TrackerMetrics};
     use crate::core::services::tracker_factory;
