+++ src/servers/http/v1/services/scrape.rs
@@ -80,7 +80,7 @@ mod tests {
     use torrust_tracker_primitives::{peer, DurationSinceUnixEpoch};
     use torrust_tracker_test_helpers::configuration;
 
-    use crate::bootstrap::app::initialize_tracker_dependencies;
+    use crate::app_test::initialize_tracker_dependencies;
     use crate::core::services::tracker_factory;
     use crate::core::Tracker;
